Started work on Meal service.
Added a Post mapping of the diary page.

Added the Entity and a table for Meals.

Added a Meal model with an Entity and table.

Added a one to many list in the User model.

Added a Dairy and Meal repository.

Added a Diaries and Meals in the services directory.